Use in process java compilation unless jdk is explicitly set by the developer
This PR updates KCT to use in process compilation unless the developer
explicitly provides a jdkHome that does not match the java home of the
current process.

In most cases, it won't be provided and because we have a default,
current implementation makes KCT spin off another process to compile java
sources, causing a significant compilation time penalty.
https://github.com/tschuchortdev/kotlin-compile-testing/issues/113#issuecomment-781731252

With this change, KCT won't create another process unless jdkHome is
set to a value that does not match the jdk home inherited from the
current process.

Issue: #113
Test: As this is a bit difficult to test, I've added test that will
check the logs ¯\_(ツ)_/¯